requirements.txt
torch>=1.12.0
pandas>=1.3.0
ultralytics>=8.0
torchvision>=0.12.0
Pillow>=8.0
matplotlib>=3.4.0
keras>=2.11.0
tensorflow>=2.11.0
numpy>=1.21.0